Switch to crypto/rand and add int() rand functions
This moves the random functions to utilize `crypto/rand` for
cryptographic safe pseudo-random. These go APIs can be a little unruly
since they use `math/big` and require some type coersions that aren't
ideal, but shouldn't be an issue since the functions utilize `int` and
not `int64` types. This adds the following functions:

- `RandPositiveInt` which generates a positive integer in half-open
  interval to match the other functions.
- `RandIntRange` which  generates a integer between `[min,max)`

Two other things that might come up with this; adding panics seems bad
(but the `rand.Intn` utilizes these as well just internally) and this
will break deterministic tests (though the current `math/rand` don't
utilize seeds in the tests anyway so this may be a non-issue for now).