Add GTFOBins reverse payloads
This adds the NodeJS, Perl, Ruby, and socat payloads from the gtfobins
project (with some modification). It's also the first steps to starting
to figure out how to best break out payload logic.

This first set of payloads were the ones that allowed for direct
inlining (ie not shelling out to python).