Adds "local" payload type and privesclib
This adds the "local" package for use with exploit specific use cases
that have appeared multiple times. The first payload in this contains a
shared object library for linux that clocks in at 204 bytes and handles
privilege escalation when run as root. This has come up a few times and
could be potentially useful elsewhere.

A note, this uses both "embed" to get the binary into the payload
variable directly and "go:generate" on the local/privesclib_amd64.asm
file in order to allow for dynamic generation.